GUI/requirements.txt
pillow==10.0.1
pyserial